Fix #214: Leave temp dir before raising exceptions
Change out of temporary directory before raising an exception that
causes the directory to be deleted as the context manager exits. If not
done, a PermissionError is raised repeatedly on Windows and the
message from the original exception is very difficult to read. This
behaviour was observed for the parmchk2 command, but the fix has also
been applied to the antechamber command.